docker-compose/tpch_workload_executor/requirements.txt
cx-Oracle==7.3.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.2
mysql==0.0.3
mysql-connector-python==8.0.18
numpy==1.18.0
pandas==0.25.3
protobuf==3.11.2
psycopg2-binary==2.8.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
